+++ sdk/securityinsight/arm-securityinsight/src/securityInsights.ts
@@ -41,6 +41,9 @@ import {
   ProductSettingsImpl,
   SourceControlOperationsImpl,
   SourceControlsImpl,
+  GetRecommendationsImpl,
+  GetImpl,
+  UpdateImpl,
   ThreatIntelligenceIndicatorImpl,
   ThreatIntelligenceIndicatorsImpl,
   ThreatIntelligenceIndicatorMetricsImpl,
@@ -77,6 +80,9 @@ import {
   ProductSettings,
   SourceControlOperations,
   SourceControls,
+  GetRecommendations,
+  Get,
+  Update,
   ThreatIntelligenceIndicator,
   ThreatIntelligenceIndicators,
   ThreatIntelligenceIndicatorMetrics,
@@ -174,7 +180,7 @@ export class SecurityInsights extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2022-09-01-preview";
+    this.apiVersion = options.apiVersion || "2022-11-01-preview";
     this.alertRules = new AlertRulesImpl(this);
     this.actions = new ActionsImpl(this);
     this.alertRuleTemplates = new AlertRuleTemplatesImpl(this);
@@ -203,6 +209,9 @@ export class SecurityInsights extends coreClient.ServiceClient {
     this.productSettings = new ProductSettingsImpl(this);
     this.sourceControlOperations = new SourceControlOperationsImpl(this);
     this.sourceControls = new SourceControlsImpl(this);
+    this.getRecommendations = new GetRecommendationsImpl(this);
+    this.get = new GetImpl(this);
+    this.update = new UpdateImpl(this);
     this.threatIntelligenceIndicator = new ThreatIntelligenceIndicatorImpl(
       this
     );
@@ -276,6 +285,9 @@ export class SecurityInsights extends coreClient.ServiceClient {
   productSettings: ProductSettings;
   sourceControlOperations: SourceControlOperations;
   sourceControls: SourceControls;
+  getRecommendations: GetRecommendations;
+  get: Get;
+  update: Update;
   threatIntelligenceIndicator: ThreatIntelligenceIndicator;
   threatIntelligenceIndicators: ThreatIntelligenceIndicators;
   threatIntelligenceIndicatorMetrics: ThreatIntelligenceIndicatorMetrics;